started working on the front end of the app

+++ resources/views/admin/products/edit_product.blade.php
@@ -74,7 +74,10 @@
                 <div class="controls">
                   <input type="file" name="image" id="image">
                   <input type="hidden" name="current_image" value="{{ $productDetails->image }}">
-                  <img style="width: 50px;" src="{{ asset('/images/backend_images/products/small/'.$productDetails->image) }}"> | <a href="/admin/delete-product-image/{{ $productDetails->id }}">Delete</a>
+                  @if(!empty($productDetails->image))
+                    <img style="width: 50px;" src="{{ asset('/images/backend_images/products/small/'.$productDetails->image) }}"> | <a href="/admin/delete-product-image/{{ $productDetails->id }}">Delete</a>
+                  @endif
+                  
                 </div>
               </div>
               <div class="form-actions">
